v 1.0.13 Pre-release. Changed market item height. Fixed work with images and incorrect exif. Extract string resources. Autolocation setup. UI and crash fixes

+++ app/src/main/java/com/ubcoin/model/response/profile/Wallet.kt
@@ -7,6 +7,6 @@ import java.io.Serializable
  */
 class Wallet(
         val number: Long?,
-        val amountUBC: Long?,
-        val amountOnHold: Long?
+        val amountUBC: Double?,
+        val amountOnHold: Double?
 ) : Serializable
